user profile

+++ src/controllers/movieController.js
@@ -1,4 +1,5 @@
-import Movie from "../models/Movie.js";
+import Movie from "../models/Movie";
+import User from "../models/User";
 
 // queries (read-list/search/watch)
 export const listMovie = async (req, res) => {
@@ -22,11 +23,11 @@ export const searchMovie = async (req, res) => {
 
 export const watchMovie = async (req, res) => {
   const { id } = req.params;
-  const movie = await Movie.findById(id);
+  const movie = await Movie.findById(id).populate("owner");
   if (!movie) {
     return res.render("404", { pageTitle: "Movie not found." });
   }
-  return res.render("detail", { pageTitle: movie.title, movie });
+  return res.render("watch", { pageTitle: movie.title, movie });
 };
 
 // mutations (create, update, delete)
@@ -36,22 +37,25 @@ export const addMovie = (req, res) => {
 export const createMovie = async (req, res) => {
   const {
     body: { title, summary, year, genres, posterImage },
+    session: { user },
     file,
   } = req;
-  console.log(`# addMovie: ${JSON.stringify(file)}`);
-
   try {
-    if (!title || !summary) {
+    if (!title || !summary || !year || !genres) {
       throw new Error("Mandatory fields are required.");
     }
-    await Movie.create({
+    const newMovie = await Movie.create({
       title,
       summary,
       year: Movie.formatYear(year),
       genres: Movie.formatGenres(genres),
       posterImage,
       fileUrl: file ? file.path : undefined,
+      owner: user._id,
     });
+    const userInfo = await User.findById(user._id);
+    userInfo.movies.push(newMovie._id);
+    userInfo.save();
     return res.redirect("/");
   } catch (error) {
     console.error("Error adding movie:", error);
@@ -64,24 +68,30 @@ export const createMovie = async (req, res) => {
 
 export const editMovie = async (req, res) => {
   const { id } = req.params;
+  const { user } = req.session;
   const movie = await Movie.findById(id);
   if (!movie) {
     return res.render("404", { pageTitle: "Movie not found." });
   }
-  return res.render("edit", { pageTitle: `Edit: ${movie.title}`, movie });
+  if (String(movie.owner) !== String(user._id)) {
+    return res.status(403).redirect("/");
+  }
+  return res.render("edit", { pageTitle: `Edit Movie`, movie });
 };
 export const updateMovie = async (req, res) => {
   const { id } = req.params;
+  const { user } = req.session;
   const movie = await Movie.exists({ _id: id });
   if (!movie) {
     return res.render("404", { pageTitle: "Movie not found." });
+  }
+  if (String(movie.owner) !== String(user._id)) {
+    return res.status(403).redirect("/");
   }
   const {
     body: { title, summary, year, rating, genres, posterImage },
     file,
   } = req;
-  console.log(`# updateMovie: ${file}`);
-
   await Movie.findByIdAndUpdate(id, {
     title,
     summary,
@@ -97,6 +107,14 @@ export const updateMovie = async (req, res) => {
 
 export const deleteMovie = async (req, res) => {
   const { id } = req.params;
+  const { user } = req.session;
+  const movie = await Movie.findById(id);
+  if (!movie) {
+    return res.status(404).render("404", { pageTitle: "Movie not found." });
+  }
+  if (String(movie.owner) !== String(user._id)) {
+    return res.status(403).redirect("/");
+  }
   await Movie.findByIdAndDelete(id);
   return res.redirect("/");
 };
